Refactor & use fixed re-render
